[WFLY-16634] Prevent deployment failures with test classes that are meant to be disabled when the SM is present
Note that this involved some weird refactoring because if the test isn't @RunAsClient, Arq seems to package the test class in the archive even if the @Deployment method doesn't specify it. So for HibernateSearchOutboxPollingTestCase and HibernateSearchV5MigrationHelperTestCase I had to refactor to get it so those test classes wouldn't fail on deployment.
