+++ testsuite/preview/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/backend/lucene/simple/HibernateSearchLuceneSimpleTestCase.java
@@ -87,6 +87,12 @@ public class HibernateSearchLuceneSimpleTestCase {
 
     @Deployment
     public static Archive<?> deploy() throws Exception {
+
+        // TODO maybe just use managed=false and deploy in the @BeforeClass / undeploy in an @AfterClass
+        if (AssumeTestGroupUtil.isSecurityManagerEnabled()) {
+            return AssumeTestGroupUtil.emptyJar(JAR_ARCHIVE_NAME);
+        }
+
         JavaArchive jar = ShrinkWrap.create(JavaArchive.class, JAR_ARCHIVE_NAME);
         // add Jakarta Persistence configuration
         jar.addAsManifestResource(HibernateSearchLuceneSimpleTestCase.class.getPackage(), "persistence.xml", "persistence.xml");
